muser: use the correct PCI_COMMAND interrupt disable(ID) bit
When PCI_COMMAND interrupt disable bit set to 1, it means
disable the pin-based INTx interrupt.

For NVMe controller, it will use INTx first, and then switch
to MSIX, the interrupt vector 0 is both valid for INTx and
MSIX, so here we remove the check when posting an interrupt,
just check the interrupt fd is valid or not.

Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>